Revert accidental change in incorrect file
