+++ src/client/pages/Bodem/config.ts
@@ -1,20 +1,13 @@
 import { generatePath } from 'react-router-dom';
 
 import { LoodMetingFrontend } from '../../../server/services/bodem/types';
-import { VarenFrontend } from '../../../server/services/varen/config-and-types';
 import { AppRoutes } from '../../../universal/config/routes';
 import { dateSort } from '../../../universal/helpers/date';
 import { LinkProps } from '../../../universal/types';
-import {
-  DisplayProps,
-  WithDetailLinkComponent,
-} from '../../components/Table/TableV2';
 
-export const displayPropsAanvragen: DisplayProps<
-  WithDetailLinkComponent<VarenFrontend>
-> = {
+export const displayPropsAanvragen = {
   detailLinkComponent: 'Adres',
-  dateRequestFormatted: 'Aangevraagd',
+  datumAanvraagFormatted: 'Aangevraagd',
   status: 'Status',
 };
