+++ src/client/pages/Varen/Varen-thema-config.ts
@@ -4,8 +4,14 @@ import { VarenFrontend } from '../../../server/services/varen/config-and-types';
 import { AppRoutes } from '../../../universal/config/routes';
 import { dateSort } from '../../../universal/helpers/date';
 import { LinkProps } from '../../../universal/types';
+import {
+  DisplayProps,
+  WithDetailLinkComponent,
+} from '../../components/Table/TableV2';
 
-export const displayPropsAanvragen = {
+export const displayPropsAanvragen: DisplayProps<
+  WithDetailLinkComponent<VarenFrontend>
+> = {
   detailLinkComponent: '',
   title: 'Omschrijving',
   dateRequestFormatted: 'Aangevraagd',
